HF/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
python-multipart==0.0.20
openai==1.58.1
pymupdf==1.25.1
tiktoken==0.8.0
faiss-cpu==1.9.0.post1
numpy==1.26.4
python-dotenv==1.0.1
